requirements.txt
python-telegram-bot>=20.0,<21.0
supabase>=2.0,<3.0
python-dotenv==1.0.0
asyncio-throttle==1.0.2
validators==0.20.0
re2==0.2.24
aiohttp>=3.8.0
